+++ base/nvti.c
@@ -97,7 +97,7 @@ vtref_new (const gchar *type, const gchar *ref_id, const gchar *ref_text)
   if (ref_text)
     ref->ref_text = g_strdup (ref_text);
 
-  return (ref);
+  return ref;
 }
 
 /**
@@ -128,7 +128,7 @@ vtref_free (vtref_t *ref)
 const gchar *
 vtref_type (const vtref_t *r)
 {
-  return (r ? r->type : NULL);
+  return r ? r->type : NULL;
 }
 
 /**
@@ -142,7 +142,7 @@ vtref_type (const vtref_t *r)
 const gchar *
 vtref_id (const vtref_t *r)
 {
-  return (r ? r->ref_id : NULL);
+  return r ? r->ref_id : NULL;
 }
 
 /**
@@ -156,7 +156,7 @@ vtref_id (const vtref_t *r)
 const gchar *
 vtref_text (const vtref_t *r)
 {
-  return (r ? r->ref_text : NULL);
+  return r ? r->ref_text : NULL;
 }
 
 /* Support function for timestamps */
@@ -209,7 +209,7 @@ parse_nvt_timestamp (const gchar *str_time)
                                 &tm)
                       == NULL)
                     {
-                      g_warning ("%s: Failed to parse time: %s", __FUNCTION__,
+                      g_warning ("%s: Failed to parse time: %s", __func__,
                                  str_time);
                       return 0;
                     }
@@ -220,7 +220,7 @@ parse_nvt_timestamp (const gchar *str_time)
   epoch_time = mktime (&tm);
   if (epoch_time == -1)
     {
-      g_warning ("%s: Failed to make time: %s", __FUNCTION__, str_time);
+      g_warning ("%s: Failed to make time: %s", __func__, str_time);
       return 0;
     }
 
@@ -241,8 +241,7 @@ parse_nvt_timestamp (const gchar *str_time)
                   "$Date: %*s %*s %*s %*u:%*u:%*u %*u %d%*[^]]", &offset)
           != 1))
     {
-      g_warning ("%s: Failed to parse timezone offset: %s", __FUNCTION__,
-                 str_time);
+      g_warning ("%s: Failed to parse timezone offset: %s", __func__, str_time);
       return 0;
     }
 
@@ -280,8 +279,9 @@ typedef struct nvti
   time_t creation_time;     /**< @brief Time of creation, seconds since epoch */
   time_t modification_time; /**< @brief Time of last change, sec. since epoch */
 
-  gchar *solution;      /**< @brief The solution */
-  gchar *solution_type; /**< @brief The solution type */
+  gchar *solution;        /**< @brief The solution */
+  gchar *solution_type;   /**< @brief The solution type */
+  gchar *solution_method; /**< @brief The solution method */
 
   gchar *tag;       /**< @brief List of tags attached to this NVT */
   gchar *cvss_base; /**< @brief CVSS base score for this NVT. */
@@ -319,10 +319,10 @@ int
 nvti_add_vtref (nvti_t *vt, vtref_t *ref)
 {
   if (!vt)
-    return (-1);
+    return -1;
 
   vt->refs = g_slist_append (vt->refs, ref);
-  return (0);
+  return 0;
 }
 
 /* VT preferences */
@@ -366,7 +366,7 @@ nvtpref_new (int id, gchar *name, gchar *type, gchar *dflt)
   if (dflt)
     np->dflt = g_strdup (dflt);
 
-  return (np);
+  return np;
 }
 
 /**
@@ -411,7 +411,7 @@ nvtpref_id (const nvtpref_t *np)
 gchar *
 nvtpref_name (const nvtpref_t *np)
 {
-  return (np ? np->name : NULL);
+  return np ? np->name : NULL;
 }
 
 /**
@@ -425,7 +425,7 @@ nvtpref_name (const nvtpref_t *np)
 gchar *
 nvtpref_type (const nvtpref_t *np)
 {
-  return (np ? np->type : NULL);
+  return np ? np->type : NULL;
 }
 
 /**
@@ -439,7 +439,7 @@ nvtpref_type (const nvtpref_t *np)
 gchar *
 nvtpref_default (const nvtpref_t *np)
 {
-  return (np ? np->dflt : NULL);
+  return np ? np->dflt : NULL;
 }
 
 /**
@@ -453,7 +453,7 @@ nvtpref_default (const nvtpref_t *np)
 nvti_t *
 nvti_new (void)
 {
-  return ((nvti_t *) g_malloc0 (sizeof (nvti_t)));
+  return (nvti_t *) g_malloc0 (sizeof (nvti_t));
 }
 
 /**
@@ -475,6 +475,7 @@ nvti_free (nvti_t *n)
   g_free (n->impact);
   g_free (n->solution);
   g_free (n->solution_type);
+  g_free (n->solution_method);
   g_free (n->tag);
   g_free (n->cvss_base);
   g_free (n->dependencies);
@@ -502,7 +503,7 @@ nvti_free (nvti_t *n)
 gchar *
 nvti_oid (const nvti_t *n)
 {
-  return (n ? n->oid : NULL);
+  return n ? n->oid : NULL;
 }
 
 /**
@@ -516,7 +517,7 @@ nvti_oid (const nvti_t *n)
 gchar *
 nvti_name (const nvti_t *n)
 {
-  return (n ? n->name : NULL);
+  return n ? n->name : NULL;
 }
 
 /**
@@ -530,7 +531,7 @@ nvti_name (const nvti_t *n)
 gchar *
 nvti_summary (const nvti_t *n)
 {
-  return (n ? n->summary : NULL);
+  return n ? n->summary : NULL;
 }
 
 /**
@@ -544,7 +545,7 @@ nvti_summary (const nvti_t *n)
 gchar *
 nvti_insight (const nvti_t *n)
 {
-  return (n ? n->insight : NULL);
+  return n ? n->insight : NULL;
 }
 
 /**
@@ -558,7 +559,7 @@ nvti_insight (const nvti_t *n)
 gchar *
 nvti_affected (const nvti_t *n)
 {
-  return (n ? n->affected : NULL);
+  return n ? n->affected : NULL;
 }
 
 /**
@@ -572,7 +573,7 @@ nvti_affected (const nvti_t *n)
 gchar *
 nvti_impact (const nvti_t *n)
 {
-  return (n ? n->impact : NULL);
+  return n ? n->impact : NULL;
 }
 
 /**
@@ -586,7 +587,7 @@ nvti_impact (const nvti_t *n)
 time_t
 nvti_creation_time (const nvti_t *n)
 {
-  return (n ? n->creation_time : 0);
+  return n ? n->creation_time : 0;
 }
 
 /**
@@ -600,7 +601,7 @@ nvti_creation_time (const nvti_t *n)
 time_t
 nvti_modification_time (const nvti_t *n)
 {
-  return (n ? n->modification_time : 0);
+  return n ? n->modification_time : 0;
 }
 
 /**
@@ -613,7 +614,7 @@ nvti_modification_time (const nvti_t *n)
 guint
 nvti_vtref_len (const nvti_t *n)
 {
-  return (n ? g_slist_length (n->refs) : 0);
+  return n ? g_slist_length (n->refs) : 0;
 }
 
 /**
@@ -628,7 +629,7 @@ nvti_vtref_len (const nvti_t *n)
 vtref_t *
 nvti_vtref (const nvti_t *n, guint p)
 {
-  return (n ? g_slist_nth_data (n->refs, p) : NULL);
+  return n ? g_slist_nth_data (n->refs, p) : NULL;
 }
 
 /**
@@ -664,7 +665,7 @@ nvti_refs (const nvti_t *n, const gchar *type, const gchar *exclude_types,
   gchar **exclude_split;
 
   if (!n)
-    return (NULL);
+    return NULL;
 
   refs = NULL;
   refs2 = NULL;
@@ -718,7 +719,7 @@ nvti_refs (const nvti_t *n, const gchar *type, const gchar *exclude_types,
 
   g_strfreev (exclude_split);
 
-  return (refs);
+  return refs;
 }
 
 /**
@@ -732,7 +733,7 @@ nvti_refs (const nvti_t *n, const gchar *type, const gchar *exclude_types,
 gchar *
 nvti_solution (const nvti_t *n)
 {
-  return (n ? n->solution : NULL);
+  return n ? n->solution : NULL;
 }
 
 /**
@@ -746,7 +747,21 @@ nvti_solution (const nvti_t *n)
 gchar *
 nvti_solution_type (const nvti_t *n)
 {
-  return (n ? n->solution_type : NULL);
+  return n ? n->solution_type : NULL;
+}
+
+/**
+ * @brief Get the solution method.
+ *
+ * @param n The NVT Info structure of which the solution method should
+ *          be returned.
+ *
+ * @return The solution method string. Don't free this.
+ */
+gchar *
+nvti_solution_method (const nvti_t *n)
+{
+  return n ? n->solution_method : NULL;
 }
 
 /**
@@ -760,7 +775,45 @@ nvti_solution_type (const nvti_t *n)
 gchar *
 nvti_tag (const nvti_t *n)
 {
-  return (n ? n->tag : NULL);
+  return n ? n->tag : NULL;
+}
+
+/**
+ * @brief Get a tag value by a tag name.
+ *
+ * @param n The NVT Info structure from where to search for the tag name.
+ *
+ * @param name The name of the tag for which to return the value.
+ *
+ * @return The tag value string as a copy or NULL if not found.
+ *         Needs to be free'd.
+ */
+gchar *
+nvti_get_tag (const nvti_t *n, const gchar *name)
+{
+  gchar **split, **point;
+
+  if (!n || n->tag == NULL || !name)
+    return NULL;
+
+  split = g_strsplit (n->tag, "|", 0);
+  point = split;
+
+  while (*point)
+    {
+      if ((strlen (*point) > strlen (name))
+          && (strncmp (*point, name, strlen (name)) == 0)
+          && ((*point)[strlen (name)] == '='))
+        {
+          gchar *ret;
+          ret = g_strdup (*point + strlen (name) + 1);
+          g_strfreev (split);
+          return ret;
+        }
+      point++;
+    }
+  g_strfreev (split);
+  return NULL;
 }
 
 /**
@@ -774,7 +827,7 @@ nvti_tag (const nvti_t *n)
 gchar *
 nvti_cvss_base (const nvti_t *n)
 {
-  return (n ? n->cvss_base : NULL);
+  return n ? n->cvss_base : NULL;
 }
 
 /**
@@ -788,7 +841,7 @@ nvti_cvss_base (const nvti_t *n)
 gchar *
 nvti_dependencies (const nvti_t *n)
 {
-  return (n ? n->dependencies : NULL);
+  return n ? n->dependencies : NULL;
 }
 
 /**
@@ -802,7 +855,7 @@ nvti_dependencies (const nvti_t *n)
 gchar *
 nvti_required_keys (const nvti_t *n)
 {
-  return (n ? n->required_keys : NULL);
+  return n ? n->required_keys : NULL;
 }
 
 /**
@@ -816,7 +869,7 @@ nvti_required_keys (const nvti_t *n)
 gchar *
 nvti_mandatory_keys (const nvti_t *n)
 {
-  return (n ? n->mandatory_keys : NULL);
+  return n ? n->mandatory_keys : NULL;
 }
 
 /**
@@ -830,7 +883,7 @@ nvti_mandatory_keys (const nvti_t *n)
 gchar *
 nvti_excluded_keys (const nvti_t *n)
 {
-  return (n ? n->excluded_keys : NULL);
+  return n ? n->excluded_keys : NULL;
 }
 
 /**
@@ -844,7 +897,7 @@ nvti_excluded_keys (const nvti_t *n)
 gchar *
 nvti_required_ports (const nvti_t *n)
 {
-  return (n ? n->required_ports : NULL);
+  return n ? n->required_ports : NULL;
 }
 
 /**
@@ -858,7 +911,7 @@ nvti_required_ports (const nvti_t *n)
 gchar *
 nvti_required_udp_ports (const nvti_t *n)
 {
-  return (n ? n->required_udp_ports : NULL);
+  return n ? n->required_udp_ports : NULL;
 }
 
 /**
@@ -872,7 +925,7 @@ nvti_required_udp_ports (const nvti_t *n)
 gchar *
 nvti_detection (const nvti_t *n)
 {
-  return (n ? n->detection : NULL);
+  return n ? n->detection : NULL;
 }
 
 /**
@@ -886,7 +939,7 @@ nvti_detection (const nvti_t *n)
 gchar *
 nvti_qod_type (const nvti_t *n)
 {
-  return (n ? n->qod_type : NULL);
+  return n ? n->qod_type : NULL;
 }
 
 /**
@@ -900,7 +953,7 @@ nvti_qod_type (const nvti_t *n)
 gchar *
 nvti_family (const nvti_t *n)
 {
-  return (n ? n->family : NULL);
+  return n ? n->family : NULL;
 }
 
 /**
@@ -913,7 +966,7 @@ nvti_family (const nvti_t *n)
 guint
 nvti_pref_len (const nvti_t *n)
 {
-  return (n ? g_slist_length (n->prefs) : 0);
+  return n ? g_slist_length (n->prefs) : 0;
 }
 
 /**
@@ -928,7 +981,7 @@ nvti_pref_len (const nvti_t *n)
 const nvtpref_t *
 nvti_pref (const nvti_t *n, guint p)
 {
-  return (n ? g_slist_nth_data (n->prefs, p) : NULL);
+  return n ? g_slist_nth_data (n->prefs, p) : NULL;
 }
 
 /**
@@ -942,7 +995,7 @@ nvti_pref (const nvti_t *n, guint p)
 gint
 nvti_timeout (const nvti_t *n)
 {
-  return (n ? n->timeout : -1);
+  return n ? n->timeout : -1;
 }
 
 /**
@@ -955,7 +1008,7 @@ nvti_timeout (const nvti_t *n)
 gint
 nvti_category (const nvti_t *n)
 {
-  return (n ? n->category : -1);
+  return n ? n->category : -1;
 }
 
 /**
@@ -971,12 +1024,11 @@ int
 nvti_set_oid (nvti_t *n, const gchar *oid)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->oid)
-    g_free (n->oid);
+  g_free (n->oid);
   n->oid = g_strdup (oid);
-  return (0);
+  return 0;
 }
 
 /**
@@ -992,12 +1044,11 @@ int
 nvti_set_name (nvti_t *n, const gchar *name)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->name)
-    g_free (n->name);
+  g_free (n->name);
   n->name = g_strdup (name);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1013,12 +1064,11 @@ int
 nvti_set_summary (nvti_t *n, const gchar *summary)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->summary)
-    g_free (n->summary);
+  g_free (n->summary);
   n->summary = g_strdup (summary);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1034,12 +1084,11 @@ int
 nvti_set_insight (nvti_t *n, const gchar *insight)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->insight)
-    g_free (n->insight);
+  g_free (n->insight);
   n->insight = g_strdup (insight);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1055,12 +1104,11 @@ int
 nvti_set_affected (nvti_t *n, const gchar *affected)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->affected)
-    g_free (n->affected);
+  g_free (n->affected);
   n->affected = g_strdup (affected);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1076,12 +1124,11 @@ int
 nvti_set_impact (nvti_t *n, const gchar *impact)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->impact)
-    g_free (n->impact);
+  g_free (n->impact);
   n->impact = g_strdup (impact);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1097,10 +1144,10 @@ int
 nvti_set_creation_time (nvti_t *n, const time_t creation_time)
 {
   if (!n)
-    return (-1);
+    return -1;
 
   n->creation_time = creation_time;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1116,10 +1163,10 @@ int
 nvti_set_modification_time (nvti_t *n, const time_t modification_time)
 {
   if (!n)
-    return (-1);
+    return -1;
 
   n->modification_time = modification_time;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1135,12 +1182,11 @@ int
 nvti_set_solution (nvti_t *n, const gchar *solution)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->solution)
-    g_free (n->solution);
+  g_free (n->solution);
   n->solution = g_strdup (solution);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1157,12 +1203,32 @@ int
 nvti_set_solution_type (nvti_t *n, const gchar *solution_type)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->solution_type)
-    g_free (n->solution_type);
+  g_free (n->solution_type);
   n->solution_type = g_strdup (solution_type);
-  return (0);
+  return 0;
+}
+
+/**
+ * @brief Set the solution method of a NVT.
+ *
+ * @param n The NVT Info structure.
+ *
+ * @param solution_method The solution method to set. A copy will be created
+ *                        from this.
+ *
+ * @return 0 for success. Anything else indicates an error.
+ */
+int
+nvti_set_solution_method (nvti_t *n, const gchar *solution_method)
+{
+  if (!n)
+    return -1;
+
+  g_free (n->solution_method);
+  n->solution_method = g_strdup (solution_method);
+  return 0;
 }
 
 /**
@@ -1187,13 +1253,13 @@ nvti_add_tag (nvti_t *n, const gchar *name, const gchar *value)
   gchar *newvalue = NULL;
 
   if (!n)
-    return (-1);
+    return -1;
 
   if (!name || !name[0])
-    return (-2);
+    return -2;
 
   if (!value || !value[0])
-    return (-3);
+    return -3;
 
   if (!strcmp (name, "last_modification"))
     {
@@ -1214,7 +1280,7 @@ nvti_add_tag (nvti_t *n, const gchar *name, const gchar *value)
        * in the NASL scripts can be removed that set "cvss_base".
        * Once this happened this exception can be removed from the code.
        */
-      return (0);
+      return 0;
     }
 
   if (n->tag)
@@ -1229,10 +1295,9 @@ nvti_add_tag (nvti_t *n, const gchar *name, const gchar *value)
   else
     n->tag = g_strconcat (name, "=", newvalue ? newvalue : value, NULL);
 
-  if (newvalue)
-    g_free (newvalue);
+  g_free (newvalue);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1248,15 +1313,14 @@ int
 nvti_set_tag (nvti_t *n, const gchar *tag)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->tag)
-    g_free (n->tag);
+  g_free (n->tag);
   if (tag && tag[0])
     n->tag = g_strdup (tag);
   else
     n->tag = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1272,15 +1336,14 @@ int
 nvti_set_cvss_base (nvti_t *n, const gchar *cvss_base)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->cvss_base)
-    g_free (n->cvss_base);
+  g_free (n->cvss_base);
   if (cvss_base && cvss_base[0])
     n->cvss_base = g_strdup (cvss_base);
   else
     n->cvss_base = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1297,15 +1360,14 @@ int
 nvti_set_dependencies (nvti_t *n, const gchar *dependencies)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->dependencies)
-    g_free (n->dependencies);
+  g_free (n->dependencies);
   if (dependencies && dependencies[0])
     n->dependencies = g_strdup (dependencies);
   else
     n->dependencies = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1322,15 +1384,14 @@ int
 nvti_set_required_keys (nvti_t *n, const gchar *required_keys)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->required_keys)
-    g_free (n->required_keys);
+  g_free (n->required_keys);
   if (required_keys && required_keys[0])
     n->required_keys = g_strdup (required_keys);
   else
     n->required_keys = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1347,15 +1408,14 @@ int
 nvti_set_mandatory_keys (nvti_t *n, const gchar *mandatory_keys)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->mandatory_keys)
-    g_free (n->mandatory_keys);
+  g_free (n->mandatory_keys);
   if (mandatory_keys && mandatory_keys[0])
     n->mandatory_keys = g_strdup (mandatory_keys);
   else
     n->mandatory_keys = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1372,15 +1432,14 @@ int
 nvti_set_excluded_keys (nvti_t *n, const gchar *excluded_keys)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->excluded_keys)
-    g_free (n->excluded_keys);
+  g_free (n->excluded_keys);
   if (excluded_keys && excluded_keys[0])
     n->excluded_keys = g_strdup (excluded_keys);
   else
     n->excluded_keys = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1397,15 +1456,14 @@ int
 nvti_set_required_ports (nvti_t *n, const gchar *required_ports)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->required_ports)
-    g_free (n->required_ports);
+  g_free (n->required_ports);
   if (required_ports && required_ports[0])
     n->required_ports = g_strdup (required_ports);
   else
     n->required_ports = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1422,15 +1480,14 @@ int
 nvti_set_required_udp_ports (nvti_t *n, const gchar *required_udp_ports)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->required_udp_ports)
-    g_free (n->required_udp_ports);
+  g_free (n->required_udp_ports);
   if (required_udp_ports && required_udp_ports[0])
     n->required_udp_ports = g_strdup (required_udp_ports);
   else
     n->required_udp_ports = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1446,12 +1503,11 @@ int
 nvti_set_detection (nvti_t *n, const gchar *detection)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->detection)
-    g_free (n->detection);
+  g_free (n->detection);
   n->detection = g_strdup (detection);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1468,15 +1524,14 @@ int
 nvti_set_qod_type (nvti_t *n, const gchar *qod_type)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->qod_type)
-    g_free (n->qod_type);
+  g_free (n->qod_type);
   if (qod_type && qod_type[0])
     n->qod_type = g_strdup (qod_type);
   else
     n->qod_type = NULL;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1492,12 +1547,11 @@ int
 nvti_set_family (nvti_t *n, const gchar *family)
 {
   if (!n)
-    return (-1);
+    return -1;
 
-  if (n->family)
-    g_free (n->family);
+  g_free (n->family);
   n->family = g_strdup (family);
-  return (0);
+  return 0;
 }
 
 /**
@@ -1513,10 +1567,10 @@ int
 nvti_set_timeout (nvti_t *n, const gint timeout)
 {
   if (!n)
-    return (-1);
+    return -1;
 
   n->timeout = timeout;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1532,10 +1586,10 @@ int
 nvti_set_category (nvti_t *n, const gint category)
 {
   if (!n)
-    return (-1);
+    return -1;
 
   n->category = category;
-  return (0);
+  return 0;
 }
 
 /**
@@ -1560,10 +1614,10 @@ nvti_add_refs (nvti_t *n, const gchar *type, const gchar *ref_ids,
   gchar **split, **item;
 
   if (!n)
-    return (1);
+    return 1;
 
   if (!ref_ids)
-    return (2);
+    return 2;
 
   split = g_strsplit (ref_ids, ",", 0);
 
@@ -1593,7 +1647,7 @@ nvti_add_refs (nvti_t *n, const gchar *type, const gchar *ref_ids,
     }
   g_strfreev (split);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1611,9 +1665,9 @@ nvti_add_required_keys (nvti_t *n, const gchar *key)
   gchar *old;
 
   if (!n)
-    return (1);
+    return 1;
   if (!key)
-    return (2);
+    return 2;
 
   old = n->required_keys;
 
@@ -1625,7 +1679,7 @@ nvti_add_required_keys (nvti_t *n, const gchar *key)
   else
     n->required_keys = g_strdup (key);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1643,9 +1697,9 @@ nvti_add_mandatory_keys (nvti_t *n, const gchar *key)
   gchar *old;
 
   if (!n)
-    return (1);
+    return 1;
   if (!key)
-    return (2);
+    return 2;
 
   old = n->mandatory_keys;
 
@@ -1657,7 +1711,7 @@ nvti_add_mandatory_keys (nvti_t *n, const gchar *key)
   else
     n->mandatory_keys = g_strdup (key);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1675,9 +1729,9 @@ nvti_add_excluded_keys (nvti_t *n, const gchar *key)
   gchar *old;
 
   if (!n)
-    return (1);
+    return 1;
   if (!key)
-    return (2);
+    return 2;
 
   old = n->excluded_keys;
 
@@ -1689,7 +1743,7 @@ nvti_add_excluded_keys (nvti_t *n, const gchar *key)
   else
     n->excluded_keys = g_strdup (key);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1707,9 +1761,9 @@ nvti_add_required_ports (nvti_t *n, const gchar *port)
   gchar *old;
 
   if (!n)
-    return (1);
+    return 1;
   if (!port)
-    return (2);
+    return 2;
 
   old = n->required_ports;
 
@@ -1721,7 +1775,7 @@ nvti_add_required_ports (nvti_t *n, const gchar *port)
   else
     n->required_ports = g_strdup (port);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1739,9 +1793,9 @@ nvti_add_required_udp_ports (nvti_t *n, const gchar *port)
   gchar *old;
 
   if (!n)
-    return (1);
+    return 1;
   if (!port)
-    return (2);
+    return 2;
 
   old = n->required_udp_ports;
 
@@ -1753,7 +1807,7 @@ nvti_add_required_udp_ports (nvti_t *n, const gchar *port)
   else
     n->required_udp_ports = g_strdup (port);
 
-  return (0);
+  return 0;
 }
 
 /**
@@ -1769,10 +1823,10 @@ int
 nvti_add_pref (nvti_t *n, nvtpref_t *np)
 {
   if (!n)
-    return (-1);
+    return -1;
 
   n->prefs = g_slist_append (n->prefs, np);
-  return (0);
+  return 0;
 }
 
 /* Collections of nvtis. */
@@ -1796,7 +1850,7 @@ free_nvti_for_hash_table (gpointer nvti)
 nvtis_t *
 nvtis_new (void)
 {
-  return g_hash_table_new_full (g_str_hash, g_str_equal, NULL,
+  return g_hash_table_new_full (g_str_hash, g_str_equal, g_free,
                                 free_nvti_for_hash_table);
 }
 
@@ -1822,7 +1876,9 @@ void
 nvtis_add (nvtis_t *nvtis, nvti_t *nvti)
 {
   if (nvti)
-    g_hash_table_insert (nvtis, (gpointer) nvti_oid (nvti), (gpointer) nvti);
+    g_hash_table_insert (
+      nvtis, (gpointer) (nvti_oid (nvti) ? g_strdup (nvti_oid (nvti)) : NULL),
+      (gpointer) nvti);
 }
 
 /**