+++ base/nvti.h
@@ -105,7 +105,11 @@ nvti_solution (const nvti_t *);
 gchar *
 nvti_solution_type (const nvti_t *);
 gchar *
+nvti_solution_method (const nvti_t *);
+gchar *
 nvti_tag (const nvti_t *);
+gchar *
+nvti_get_tag (const nvti_t *, const gchar *);
 gchar *
 nvti_cvss_base (const nvti_t *);
 gchar *
@@ -155,6 +159,8 @@ int
 nvti_set_solution (nvti_t *, const gchar *);
 int
 nvti_set_solution_type (nvti_t *, const gchar *);
+int
+nvti_set_solution_method (nvti_t *, const gchar *);
 int
 nvti_add_tag (nvti_t *, const gchar *, const gchar *);
 int